Day 25 in Scratch
Scratch cannot load external files, such as the input text. Instead, create a new list named `text_input`; then right-click it, choose "Import", and select the input file. This loads the file line-by-line into the list (as shown in *day25-example.png*).
The program is represented as an image in the `.png` file and as text in the `.scratchblocks` file.